+++ KesaApplication/AfterRegisterViewController.swift
@@ -1,0 +1,24 @@
+//
+//  AfterRegisterViewController.swift
+//  KesaApplication
+//
+//  Created by Joon Lee on 2016-06-08.
+//  Copyright © 2016 kesa. All rights reserved.
+//
+
+import UIKit
+
+class AfterRegisterViewController: UIViewController {
+
+    override func viewDidLoad() {
+        super.viewDidLoad()
+
+        // Do any additional setup after loading the view.
+    }
+
+    override func didReceiveMemoryWarning() {
+        super.didReceiveMemoryWarning()
+        // Dispose of any resources that can be recreated.
+    }
+
+}
