+++ KesaApplication/Base.lproj/Main.storyboard
@@ -5,10 +5,10 @@
         <capability name="Constraints to layout margins" minToolsVersion="6.0"/>
     </dependencies>
     <scenes>
-        <!--View Controller-->
+        <!--Register View Controller-->
         <scene sceneID="tne-QT-ifu">
             <objects>
-                <viewController id="BYZ-38-t0r" customClass="ViewController" customModule="KesaApplication" customModuleProvider="target" sceneMemberID="viewController">
+                <viewController id="BYZ-38-t0r" customClass="RegisterViewController" customModule="KesaApplication" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="y3c-jy-aDJ"/>
                         <viewControllerLayoutGuide type="bottom" id="wfy-db-euE"/>
@@ -165,30 +165,45 @@
             </objects>
             <point key="canvasLocation" x="220" y="259"/>
         </scene>
-        <!--View Controller-->
+        <!--After Register View Controller-->
         <scene sceneID="EEH-Me-eGF">
             <objects>
-                <viewController id="iFM-aS-Xff" sceneMemberID="viewController">
+                <viewController id="iFM-aS-Xff" customClass="AfterRegisterViewController" customModule="KesaApplication" customModuleProvider="target" sceneMemberID="viewController">
                     <layoutGuides>
                         <viewControllerLayoutGuide type="top" id="U0Q-Br-Z4n"/>
                         <viewControllerLayoutGuide type="bottom" id="UHg-cP-UP6"/>
                     </layoutGuides>
-                    <view key="view" contentMode="scaleToFill" id="aqJ-oD-5Jf">
+                    <view key="view" contentMode="scaleToFill" id="aqJ-oD-5Jf" customClass="AfterRegisterViewContro">
                         <rect key="frame" x="0.0" y="0.0" width="600" height="600"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <button opaque="NO" contentMode="scaleToFill" fixedFrame="YES" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="1nJ-0H-eUr">
-                                <rect key="frame" x="281" y="260" width="38" height="30"/>
-                                <state key="normal" title="Login"/>
+                            <button opaque="NO" contentMode="scaleToFill" contentHorizontalAlignment="center" contentVerticalAlignment="center" buttonType="roundedRect" lineBreakMode="middleTruncation" translatesAutoresizingMaskIntoConstraints="NO" id="1nJ-0H-eUr">
+                                <rect key="frame" x="20" y="260" width="560" height="30"/>
+                                <fontDescription key="fontDescription" name="GillSans" family="Gill Sans" pointSize="17"/>
+                                <state key="normal" title="Login">
+                                    <color key="titleColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
+                                </state>
                             </button>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" fixedFrame="YES" text="You've successfully created an account!" textAlignment="natural" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="cbl-Ux-hTe">
-                                <rect key="frame" x="147" y="179" width="306" height="21"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="You've successfully created an account!" textAlignment="center" lineBreakMode="tailTruncation" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="cbl-Ux-hTe">
+                                <rect key="frame" x="20" y="181" width="560" height="21"/>
+                                <constraints>
+                                    <constraint firstAttribute="height" constant="21" id="9Vw-rk-3vf"/>
+                                </constraints>
+                                <fontDescription key="fontDescription" name="GillSans" family="Gill Sans" pointSize="19"/>
                                 <color key="textColor" red="0.0" green="0.0" blue="0.0" alpha="1" colorSpace="calibratedRGB"/>
                                 <nil key="highlightedColor"/>
                             </label>
                         </subviews>
                         <color key="backgroundColor" white="1" alpha="1" colorSpace="calibratedWhite"/>
+                        <constraints>
+                            <constraint firstItem="1nJ-0H-eUr" firstAttribute="leading" secondItem="aqJ-oD-5Jf" secondAttribute="leading" constant="20" id="43u-Lj-r7U"/>
+                            <constraint firstItem="1nJ-0H-eUr" firstAttribute="top" secondItem="cbl-Ux-hTe" secondAttribute="bottom" constant="58" id="Gea-Ra-Ubc"/>
+                            <constraint firstAttribute="trailing" secondItem="cbl-Ux-hTe" secondAttribute="trailing" constant="20" id="LaG-ZX-us5"/>
+                            <constraint firstItem="cbl-Ux-hTe" firstAttribute="leading" secondItem="aqJ-oD-5Jf" secondAttribute="leading" constant="20" id="fCp-P8-h6T"/>
+                            <constraint firstItem="cbl-Ux-hTe" firstAttribute="top" secondItem="U0Q-Br-Z4n" secondAttribute="bottom" constant="161" id="fmF-WY-rEf"/>
+                            <constraint firstItem="UHg-cP-UP6" firstAttribute="top" secondItem="1nJ-0H-eUr" secondAttribute="bottom" constant="310" id="oC2-ae-4II"/>
+                            <constraint firstAttribute="trailing" secondItem="1nJ-0H-eUr" secondAttribute="trailing" constant="20" id="x0i-Kx-fl0"/>
+                        </constraints>
                     </view>
                 </viewController>
                 <placeholder placeholderIdentifier="IBFirstResponder" id="nz0-2D-IP3" userLabel="First Responder" sceneMemberID="firstResponder"/>
